Small UI improvements
- Add Select All and Deselect All to body parts checks
- Move common parts to correct folders
