+++ animationCombiner/utils/__init__.py
@@ -1,4 +1,5 @@
 """Collection of all utility functions/classes that didnt fit anywhere else"""
+import bpy
 from bpy.types import PropertyGroup, Property, bpy_prop_collection
 
 
@@ -34,3 +35,26 @@ def copy(from_prop: Property, to_prop: Property, depth=0) -> None:
         to_prop.clear()
         for from_subprop in from_prop.values():
             copy(from_subprop, to_prop.add(), depth + 1)
+
+
+def on_actions_update(self=None, context=None):
+    """Recalculates length of final animation after the actions were updated"""
+    armature = bpy.context.view_layer.objects.active.data
+    length = 0
+    for group in armature.groups:
+        group.errors.clear()
+        group_length = 0
+        parts = set()
+        for action in group.actions:
+            group_length = max(action.length_group.length, group_length)
+            for part in action.body_parts:
+                if part.checked:
+                    if part.uuid in parts:
+                        group.add_error("COLLIDING_PARTS")
+                        break
+                    parts.add(part.uuid)
+        group.length = group_length
+        group.actions_count = len(group.actions)
+        length += group_length
+    armature.animation_length = length
+    armature.is_applied = False
